requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
jinja2>=3.1.0
python-multipart>=0.0.6
aiofiles>=23.2.1
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.7.0
litellm>=1.15.0
click>=8.1.0
